requirements.txt
scikit-learn==0.23.2
scipy==1.5.2
tensorflow-gpu==2.3.1
tensorflow-addons==0.11.2
tensorflow-datasets==4.0.1
tqdm==4.50.2
pandas==1.2.0
Pillow==8.1.0
seaborn==0.11.1